requirements.txt
Django==2.2
django-filter==2.2.0
djangorestframework==3.8.2
importlib-metadata==1.6.0
Markdown==3.2.2
mysqlclient==1.4.6
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
django-crispy-forms
django-import-export
django-reversion
django-formtools
future
httplib2
six
xlwt
xlsxwriter
requests
